// src/types.tsx
export type SearchResult = {
  login: string;
  id: number;
  node_id: string;
  avatar_url: string;
  gravatar_id?: string;
  url: string;
  html_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  starred_url: string;
  subscriptions_url: string;
  organizations_url: string;
  repos_url: string;
  events_url: string;
  received_events_url: string;
  type: string;
  user_view_type?: string;
  site_admin: boolean;
  score?: number;
};

export type UserData = {
  login: string;
  id: number;
  node_id: string;
  avatar_url: string;
  gravatar_id?: string;
  url: string;
  html_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  starred_url: string;
  subscriptions_url: string;
  organizations_url: string;
  repos_url: string;
  events_url: string;
  received_events_url: string;
  type: string;
  site_admin: boolean;
  name?: string;
  company?: string | null;
  blog?: string;
  location?: string | null;
  email?: string | null;
  hireable?: boolean | null;
  bio?: string;
  twitter_username?: string | null;
  public_repos?: number;
  public_gists?: number;
  followers?: number;
  following?: number;
  created_at?: string;
  updated_at?: string;
};

export type RepoType = {
  name: string;
  html_url: string;
  url: string;
  description?: string | null;
  forks_count: number;
  stargazers_count: number;
  license?: {
    key: string;
    name: string;
    spdx_id: string;
  } | null;
  updated_at?: string;
  [key: string]: unknown;
};